Rename utils folder to utilities to avoid file confusion on module imports. Update file paths.
