+++ api/routes/report/check-your-answers.js
@@ -4,7 +4,7 @@ import fs from 'fs';
 import path from 'path';
 var cloneDeep = require('lodash.clonedeep');
 import { azureUpload } from '../../../services';
-import { formatValidationErrors } from '../../../utils';
+import { formatValidationErrors } from '../../../utilities';
 import config from '../../../app/config';
 
 export default function (app) {
@@ -100,8 +100,7 @@ export default function (app) {
             // Push image(s) to Azure
             Object.values(req.session.data.property).forEach((item) => {
               const imageData = fs.createReadStream(
-                `${path.resolve(__dirname + '../../../../uploads/')}/${
-                  item.image
+                `${path.resolve(__dirname + '../../../../uploads/')}/${item.image
                 }`
               );
