+++ utilities/addCheckedFunction.js
