+++ utilities/forceHttps.js
