+++ utilities/index.js
@@ -4,4 +4,4 @@ export { forceHttps } from './forceHttps';
 export { formatValidationErrors } from './formatValidationErrors';
 export { matchRoutes } from './matchRoutes';
 export { sessionData } from './sessionData';
-export { GT_OSGB } from './geotools';
+export { GT_OSGB as default } from './geotools';
