+++ utilities/sessionData.js
